Add files via upload
So this is my prototype of the change I talked about. I know it has added two extra screens, but it's the one solution I've found and this is something we can discuss during our next meeting. If it is a problem, please let me know. It currently doesn't store your password, and until it does, I'm unable to check whether a user is first time or not. I'm also working on implementing a "fail safe" when a user goes to enter their password that checks to see if the field that they entered for their password is empty. However, again, almost all of this relies on the code being able to actually store the user's password.